added some missed params and methods

+++ src/main/java/org/treblereel/gwt/three4g/geometries/PlaneBufferGeometry.java
@@ -11,39 +11,39 @@ import org.treblereel.gwt.three4g.core.BufferGeometry;
 @JsType(isNative = true, namespace = "THREE")
 public class PlaneBufferGeometry extends BufferGeometry {
 
-    public int width; // — Width along the X axis. Default is 1.
+    public float width; // — Width along the X axis. Default is 1.
 
-    public int  height; // — Height along the Y axis. Default is 1.
+    public float height; // — Height along the Y axis. Default is 1.
 
-    public int  widthSegments; // — Default is 1.
+    public int widthSegments; // — Default is 1.
 
-    public int  heightSegments; // — Default is 1.
+    public int heightSegments; // — Default is 1.
 
     /**
      * @param width  — Width along the X axis. Default is 1.
      * @param height — Height along the Y axis. Default is 1.
      */
     @JsConstructor
-    public PlaneBufferGeometry(int width, int height) {
+    public PlaneBufferGeometry(float width, float height) {
     }
 
     /**
-     * @param width  — Width along the X axis. Default is 1.
-     * @param height — Height along the Y axis. Default is 1.
+     * @param width         — Width along the X axis. Default is 1.
+     * @param height        — Height along the Y axis. Default is 1.
      * @param widthSegments — Default is 1.
      */
     @JsConstructor
-    public PlaneBufferGeometry(int width, int height, int widthSegments) {
+    public PlaneBufferGeometry(float width, float height, int widthSegments) {
     }
 
     /**
-     * @param width  — Width along the X axis. Default is 1.
-     * @param height — Height along the Y axis. Default is 1.
-     * @param widthSegments — Default is 1.
+     * @param width          — Width along the X axis. Default is 1.
+     * @param height         — Height along the Y axis. Default is 1.
+     * @param widthSegments  — Default is 1.
      * @param heightSegments — Default is 1.
      */
     @JsConstructor
-    public PlaneBufferGeometry(int width, int height, int widthSegments, int heightSegments) {
+    public PlaneBufferGeometry(float width, float height, int widthSegments, int heightSegments) {
     }
 
 }
